spliting joplinCloud utils and fixing interface
