+++ packages/app-mobile/components/screens/ConfigScreen/configScreenStyles.ts
@@ -5,6 +5,7 @@ type SidebarButtonStyle = ViewStyle & { height: number };
 
 export interface ConfigScreenStyleSheet {
 	body: ViewStyle;
+	divider: TextStyle;
 
 	settingContainer: ViewStyle;
 	settingContainerNoBottomBorder: ViewStyle;
@@ -102,6 +103,9 @@ const configScreenStyles = (themeId: number): ConfigScreenStyles => {
 			flex: 1,
 			justifyContent: 'flex-start',
 			flexDirection: 'column',
+		},
+		divider: {
+			color: theme.dividerColor,
 		},
 		settingContainer: settingContainerStyle,
 		settingContainerNoBottomBorder: {
